[fix]HPBarに現在の数値が正しく反映されなかったのを修正

+++ Assets/_WorkSpace/Script/Player/PlayerHealth.cs
@@ -3,7 +3,7 @@ using UnityEngine.SceneManagement;
 
 public class PlayerHealth : MonoBehaviour
 {
-    public int _playerHealth = 100;
+    public float _playerHealth = 100;
     [SerializeField]
     int _heal = 3;
     [SerializeField]
